console: add sbi_putc_raw() for output character
when linux kernel printk a string ending with '\n', kernel will send '\r\n'
to console. but if the early prink with sbi_console, it will output another
'\r' when output '\n', so the '\n' output by kernel will be '\r\r\n', then
there will be some blank in the log file.

Change-Id: If2424e1e2e808c8d830aa5e8f630d27d6a6cffa4
Signed-off-by: zhangmeng <zhangmeng.kevin@spacemit.com>